Implemented CRUD version of REST api

+++ src/main/java/spd/trello/web/WorkspaceRESTController.java
@@ -1,13 +1,10 @@
 package spd.trello.web;
 
-import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.servlet.ModelAndView;
 import spd.trello.domain.Workspace;
-import spd.trello.domain.WorkspaceVisibility;
+import spd.trello.exception.ObjectNotFoundException;
 import spd.trello.service.WorkspaceService;
 
 import java.util.List;
@@ -26,45 +23,30 @@ public class WorkspaceRESTController {
         return service.getAll();
     }
 
-    @RequestMapping("/workspace_menu")
-    public HttpEntity<Workspace> edit(ModelAndView modelAndView) {
-        return new ResponseEntity<>(service.read((UUID) modelAndView.getModel().get("workspace_id")), HttpStatus.OK);
+    @RequestMapping(value = "/workspaces/create")
+    public ResponseEntity<Workspace> Create(@RequestBody Workspace workspace) {
+        return new ResponseEntity<>(service.create(workspace),HttpStatus.OK);
     }
 
-    @RequestMapping(value = "/workspaces/create", method = RequestMethod.GET)
-    public Object Create(Model model) {
-        model.addAttribute("workspace", new Workspace());
-        model.addAttribute("action","create");
-        return "workspace/menu";
-    }
-    @RequestMapping(value = "/workspaces/update", method = RequestMethod.GET)
-    public String Update(Model model, @RequestParam("id") String id) {
-        Workspace workspace = service.read(UUID.fromString(id));
-        model.addAttribute("workspace", workspace);
-        model.addAttribute("action","update");
-        return "workspace/menu";
+    @RequestMapping(value = "/workspaces/update")
+    public ResponseEntity Update(@RequestBody Workspace workspace) {
+        if (service.read(workspace.getId()) == null)
+            return new ResponseEntity(new ObjectNotFoundException(workspace.getId(),workspace.getClass()), HttpStatus.NOT_FOUND);
+        service.update(workspace);
+        return new ResponseEntity(service.read((workspace.getId())), HttpStatus.OK);
     }
 
-    @RequestMapping(value = {"/workspaces/create", "/workspaces/update"},method = RequestMethod.POST)
-    public String saveNew(@RequestParam("action") String action, @ModelAttribute("id") String id, @RequestParam("name") String name, @RequestParam("description") String description, @RequestParam("visibility") String visibility)
-    {
-        Workspace workspace = new Workspace();
-        workspace.setId(UUID.fromString(id));
-        workspace.setName(name);
-        workspace.setDescription(description);
-        workspace.setVisibility(WorkspaceVisibility.valueOf(visibility));
-        if (action.equals("create"))
-            service.create(workspace);
-        else if (action.equals("update"))
-            service.update(workspace);
-        return "workspace/list";
-
-    }
 
     @RequestMapping("/workspaces/delete")
-    public String Delete(@RequestParam("id") String id)
+    public HttpStatus Delete(@RequestBody String id)
     {
+
+        if (service.read(UUID.fromString(id)) == null)
+            return HttpStatus.NOT_FOUND;
         service.delete(UUID.fromString(id));
-        return "workspace/list";
+        if (service.read(UUID.fromString(id)) == null)
+            return HttpStatus.OK;
+        else
+            return HttpStatus.INTERNAL_SERVER_ERROR;
     }
 }
